requirements.txt
#Ubuntu: apt install libxml2-dev libxslt1-dev python-dev
Django>1.7,<1.7.999
django-braces==1.4.0
django-cors-headers==1.0.0
django-extensions==1.5.0
django-jsonview==0.4.3
djangorestframework==3.0.5
gunicorn==19.2.1
jdcal==1.0
lxml==3.4.2
openpyxl==2.1.5
python-dateutil==2.4.0
python-mimeparse==0.1.4
six==1.9.0
sqlparse==0.1.14
defusedxml==0.4.1
mailer==0.7